Add fsync() in test_copy_file_range() to overcome un-resolved issue with copy_file_range failure

+++ test/test_syscalls.c
@@ -748,7 +748,11 @@ static int test_copy_file_range(void)
 		close(fd_in);
 		return -1;
 	}
-
+    res = fsync(fd_in);
+	if (res == -1) {
+		PERROR("fsync");
+		return -1;
+	}
 	unlink(testfile2);
 	fd_out = creat(testfile2, 0644);
 	if (fd_out == -1) {
@@ -1647,7 +1651,7 @@ fail:
 
 	rmdir(PATH("a/d/e"));
 	rmdir(PATH("a/d"));
- 
+
  	rmdir(PATH("a/b/c"));
 	rmdir(PATH("a/b"));
 	rmdir(PATH("a"));
@@ -1833,7 +1837,7 @@ int main(int argc, char *argv[])
 	err += test_symlink();
 	err += test_link();
 	err += test_link2();
-#ifndef __FreeBSD__	
+#ifndef __FreeBSD__
 	err += test_mknod();
 	err += test_mkfifo();
 #endif
